restructure types to categorize between requests and responses
