+++ canvas_api/src/lib.rs
@@ -1,28 +1,39 @@
 pub mod types {
-    pub(crate) trait ResponseType:
-        std::fmt::Debug + std::cmp::PartialEq + serde::de::DeserializeOwned
-    {
+    pub mod requests {
+        pub(crate) trait RequestType: std::fmt::Debug + serde::ser::Serialize {}
+        impl<T: RequestType> RequestType for Vec<T> {}
+
+        mod submission;
+
+        pub use submission::*;
     }
-    impl<T: ResponseType> ResponseType for Vec<T> {}
 
-    pub(crate) trait RequestType: std::fmt::Debug + serde::ser::Serialize {}
-    impl<T: RequestType> RequestType for Vec<T> {}
+    pub mod responses {
+        pub(crate) trait ResponseType:
+            std::fmt::Debug + std::cmp::PartialEq + serde::de::DeserializeOwned
+        {
+        }
+        impl<T: ResponseType> ResponseType for Vec<T> {}
 
-    mod assignment;
-    mod course;
-    mod discussion;
-    mod module;
-    mod profile;
-    mod submission;
-    mod todo;
+        mod assignment;
+        mod course;
+        mod discussion;
+        mod module;
+        mod profile;
+        mod submission;
+        mod todo;
 
-    pub use assignment::*;
-    pub use course::*;
-    pub use discussion::*;
-    pub use module::*;
-    pub use profile::*;
-    pub use submission::*;
-    pub use todo::*;
+        pub use assignment::*;
+        pub use course::*;
+        pub use discussion::*;
+        pub use module::*;
+        pub use profile::*;
+        pub use submission::*;
+        pub use todo::*;
+    }
+
+    pub use requests::*;
+    pub use responses::*;
 }
 
 pub mod requests;
