+++ canvas_api/src/types/responses/course.rs
